+++ lib/eve_online/esi/base.rb
@@ -121,13 +121,17 @@ module EveOnline
 #        class Net::WriteTimeout; end
 #      end
 
+      if Gem::Version.new(RUBY_VERSION) < Gem::Version.new('2.6.0')
+        class Net::WriteTimeout < StandardError; end
+      end
+
       def uri
         @uri ||= URI.parse(url)
       end
 
       def resource
         @resource ||= client.request(request)
-      rescue Net::OpenTimeout, Net::ReadTimeout #, Net::WriteTimeout
+      rescue Net::OpenTimeout, Net::ReadTimeout, Net::WriteTimeout
         raise EveOnline::Exceptions::Timeout
       end